moved OAuth2Manager and HubSpotAuthError from api-auth-lib to cms-lib and removed api-auth-lib
