--- packages/api-auth-lib/CHANGELOG.md
@@ -1,16 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.0.9](https://github.com/HubSpot/hubspot-cms-tools/compare/v1.0.9-beta.7...v1.0.9) (2020-02-25)
-
-**Note:** Version bump only for package @hubspot/api-auth-lib
-
-
-
-
-
-## [1.0.5](https://github.com/HubSpot/hubspot-cms-tools/compare/v1.0.5-beta.4...v1.0.5) (2020-01-24)
-
-**Note:** Version bump only for package @hubspot/api-auth-lib
