--- packages/api-auth-lib/README.md
@@ -1,4 +1,0 @@
-# HubSpot Auth
-
-Library designed to make it easy to authenticate CLI tools with HubSpot.
-
