--- packages/api-auth-lib/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  OAuth2Manager: require('./OAuth2Manager'),
-};
