--- packages/api-auth-lib/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@hubspot/api-auth-lib",
-  "version": "1.1.3-beta.0",
-  "license": "Apache-2.0",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/HubSpot/hubspot-cms-tools"
-  },
-  "dependencies": {
-    "@hubspot/cms-lib": "^1.1.4",
-    "express": "^4.16.3",
-    "moment": "^2.24.0",
-    "open": "^6.4.0",
-    "request": "^2.88.0",
-    "request-promise-native": "^1.0.7"
-  },
-  "engines": {
-    "node": ">=8.9.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "0659fd19cabc3645af431b177c11d0c1b089e0f8"
-}
