+++ packages/cms-lib/errorHandlers.js
@@ -1,4 +1,4 @@
-const { HubSpotAuthError } = require('@hubspot/api-auth-lib/Errors');
+const { HubSpotAuthError } = require('./lib/Errors');
 const { logger } = require('./logger');
 
 const isApiStatusCodeError = err =>
