+++ packages/cms-lib/lib/OAuth2Manager.js
@@ -4,13 +4,10 @@ const moment = require('moment');
 const open = require('open');
 
 const { HubSpotAuthError } = require('./Errors');
-const { ENVIRONMENTS } = require('@hubspot/cms-lib/lib/constants');
-const { handleExit } = require('@hubspot/cms-lib/lib/process');
-const {
-  getHubSpotWebsiteOrigin,
-  getHubSpotApiOrigin,
-} = require('@hubspot/cms-lib/lib/urls');
-const { getValidEnv } = require('@hubspot/cms-lib/lib/environment');
+const { ENVIRONMENTS } = require('./constants');
+const { handleExit } = require('./process');
+const { getHubSpotWebsiteOrigin, getHubSpotApiOrigin } = require('./urls');
+const { getValidEnv } = require('./environment');
 
 const PORT = 3000;
 const redirectUri = `http://localhost:${PORT}/oauth-callback`;
