+++ packages/cms-lib/package.json
@@ -13,12 +13,14 @@
     "chokidar": "^3.0.1",
     "content-disposition": "^0.5.3",
     "debounce": "^1.2.0",
+    "express": "4.16.3",
     "extract-zip": "^1.6.7",
     "findup-sync": "^3.0.0",
     "fs-extra": "^8.1.0",
     "ignore": "^5.1.4",
     "js-yaml": "^3.12.2",
     "moment": "^2.24.0",
+    "open": "6.4.0",
     "p-queue": "^6.0.2",
     "prettier": "^1.19.1",
     "request": "^2.87.0",
